requirements.txt
aiofiles==23.1.0
aiohttp==3.8.4
aiosignal==1.3.1
altair==4.2.2
anyio==3.6.2
async-timeout==4.0.2
attrs==22.2.0
brotlipy==0.7.0
cachetools==5.2.0
certifi==2020.12.5
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
conda==4.9.2
contourpy==1.0.7
csbaoyan==0.0.1
cycler==0.11.0
entrypoints==0.4
fastapi==0.94.0
ffmpy==0.3.0
fonttools==4.39.0
frozenlist==1.3.3
fsspec==2023.3.0
google-api-core==2.8.2
google-api-python-client==2.57.0
google-auth==2.10.0
google-auth-httplib2==0.1.0
googleapis-common-protos==1.56.4
gradio==3.20.1
h11==0.14.0
httpcore==0.16.3
httplib2==0.20.4
httpx==0.23.3
importlib-resources==5.12.0
Jinja2==3.1.2
jsonschema==4.17.3
jwt==1.3.1
kiwisolver==1.4.4
linkify-it-py==2.0.0
markdown-it-py==2.2.0
MarkupSafe==2.1.2
matplotlib==3.7.1
mdit-py-plugins==0.3.3
mdurl==0.1.2
menuinst==1.4.16
multidict==6.0.4
numpy==1.23.1
orjson==3.8.7
packaging==23.0
pandas==1.4.3
Pillow==9.4.0
protobuf==4.21.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycosat==0.6.3
pycryptodome==3.17
pydantic==1.10.6
pydotplus==2.0.2
pydub==0.25.1
pyparsing==3.0.9
pyrsistent==0.19.3
python-dateutil==2.8.2
python-multipart==0.0.6
pytz==2022.1
pywin32==228
PyYAML==6.0
qcloudsms-py==0.1.4
rfc3986==1.5.0
rsa==4.9
sklearn==0.0.post1
sniffio==1.3.0
starlette==0.26.0.post1
toolz==0.12.0
typing-extensions==4.5.0
uc-micro-py==1.0.1
uritemplate==4.1.1
uvicorn==0.21.0
websockets==10.4
wincertstore==0.2
xlrd==2.0.1
yarl==1.8.2
zipp==3.15.0
